+++ src/screens/app/Favorites/index.js
@@ -1,11 +1,13 @@
 
 import React, { Component } from 'react';
-import { View, Text, StatusBar, FlatList, ActivityIndicator } from 'react-native';
+import { View, Text, StatusBar, FlatList, RefreshControl } from 'react-native';
 import { Content, Container } from 'native-base';
 import { getGenres, showGenresByPage, searcher } from '../../../api/home'
 import { connect } from 'react-redux';
 //customs
 import styles from './styles';
+import FavoriteCard from '../../../components/FavoriteCard';
+import NoImage from '../../../assets/images/NoImageAvailable.png';
 import colors from '../../../utils/colors';
 import TopTabBar from '../../../components/HeaderTabs';
 import Header from '../../../components/Header';
@@ -15,65 +17,32 @@ class Favorites extends Component {
     state = {
         urlNextPage: null,
         isLoading: false,
-        categories: [],
-        indexTab: 0,
         querySearch: {
             value: '',
             type: 'text',
             required: 'false'
         },
         isLoadingNext: false,
+        favorites: [],
+        refreshing: false,
     }
-    // componentDidMount = async () => {
-    //     this.getCategories();
-    // };
 
-    // getCategories = async (isRefresh = false) => {
-    //     this.setState({
-    //         isLoading: true
-    //     });
-    //     try {
-    //         const getGenresList = await getGenres();
-    //         console.log(getGenresList, 'getGenresList')
-    //         this.setState({
-    //             categories: getGenresList.data.data,
-    //             urlNextPage: getGenresList.data.links.next,
-    //             isLoading: false,
-    //         })
-    //     } catch (err) {
-    //         console.log(err)
-    //         this.setState({
-    //             isLoading: false
-    //         });
-    //     }
-    // };
-
-    // loadMoreData = async () => {
-    //     this.setState({
-    //         isLoadingNext: true
-    //     })
-    //     try {
-    //         let { urlNextPage, categories, isLoadingNext } = this.state;
-    //         if (urlNextPage) {
-    //             let interArray = [...categories];
-    //             let response = await showGenresByPage(urlNextPage);
-    //             let nextArray = [...interArray, ...response.data.data];
-    //             this.setState({
-    //                 categories: nextArray,
-    //                 urlNextPage: response.data.links.next,
-    //             });
-    //         }
-    //         this.setState({
-    //             isLoadingNext: false,
-    //         })
-    //     } catch (err) {
-    //         console.log(err);
-    //         this.setState({
-    //             isLoadingNext: false
-    //         });
-    //     }
-    // };
-
+    componentDidMount = () => {
+        this.getFavorites();
+    }
+    getFavorites = () => {
+        this.setState({
+            refreshing: true
+        })
+        let { favorites } = this.props;
+        let removeduplicated = [...new Set(favorites)]
+        this.setState({
+            favorites: removeduplicated
+        })
+        this.setState({
+            refreshing: false
+        })
+    };
     onChangeInputValue = async (key, value) => {
         let { state } = this;
         state[key].value = value;
@@ -81,8 +50,17 @@ class Favorites extends Component {
         this.setState({ ...state });
     };
 
+    imageExists = image => {
+        if (image && Reflect.has(image, 'original')) {
+            return { uri: image.original };
+        } else {
+            return NoImage;
+        }
+    };
+
     render() {
-        let { } = this.state;
+        let { favorites, refreshing } = this.state;
+        console.log(favorites, 'favoritesList')
         return (
             <View style={styles.container}>
                 <StatusBar translucent backgroundColor='transparent' />
@@ -93,13 +71,36 @@ class Favorites extends Component {
                     onChangeValue={this.onChangeInputValue}
                 />
                 <Container style={styles.backgroundContainer}>
-                    <Content showsVerticalScrollIndicator={false} contentContainerStyle={styles.backgroundContent}>
-
+                    <Content
+                        showsVerticalScrollIndicator={false}
+                        contentContainerStyle={styles.backgroundContent}
+                        refreshControl={<RefreshControl
+                            refreshing={refreshing}
+                            onRefresh={() => this.getFavorites()}
+                        />}
+                    >
+                        <View>
+                        <Text style={styles.title}>{favorites.length > 0 ? 'Favorites' : 'No Favorites'}</Text>
+                            <FlatList
+                                keyExtractor={i => i.id}
+                                data={favorites}
+                                showsVerticalScrollIndicator={false}
+                                renderItem={({ item, index }) => (
+                                    <FavoriteCard
+                                        picture={this.imageExists(item.attributes.posterImage)}
+                                        animeName={item.attributes.canonicalTitle}
+                                    />)
+                                }
+                            />
+                        </View>
                     </Content>
                 </Container>
-            </View>
+            </View >
         );
     }
 }
+const mapStateToProps = state => ({
+    favorites: state.favorites.favorites,
+});
 
-export default Favorites
+export default connect(mapStateToProps)(Favorites);